backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
httpx==0.26.0
